backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
# Needed for FastAPI CORS middleware
starlette==0.37.2
